requirements.txt
fbchat_muqit
google-generativeai
pymongo
